Break ViewRefresher out from cache.go into refresher.go.
